+++ workspace/packages/common/constants.ts
@@ -64,3 +64,19 @@ export const SOLUTIONS_TEST_EXAMPLES_CONTENT = `\
 [
 ]
 `;
+
+/**
+ * 题解函数名称提取正则表达式
+ */
+export const SOLUTION_FUNCTION_NAME_REGEXP = {
+    // public int[] twoSum(int[] nums, int target) {
+    [Language.java]: /^\s*public\s+(?<return_type>\S+)\s+(?<function_name>\w+)\s*\((?<arguments>.*)\)\s*\{\s*$/m,
+    // func twoSum(nums []int, target int) []int {
+    [Language.golang]: /^\s*func\s+(?<function_name>\w+)\s*\((?<arguments>.*)\)\s+(?<return_type>\S+)\s*\{\s*$/m,
+    // const twoSum = function (nums, target) {
+    [Language.javascript]: /^\s*const\s+(?<function_name>\w+)\s*=\s*function\s*\((?<arguments>.*)\)\s*\{\s*$/m,
+    // function twoSum(nums: number[], target: number): number[] {
+    [Language.typescript]: /^\s*function\s+(?<function_name>\w+)\s*\((?<arguments>.*)\)\s*:\s*(?<return_type>\S+)\s*\{\s*?/m,
+} as const;
+
+export const SOLUTION_FUNCTION_NAME_REGEXP_MAP = new Map(Object.entries(SOLUTION_FUNCTION_NAME_REGEXP)) as Map<Language, RegExp>;